# config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "Talent Management System",
        "description": "TMS REST APIs based on Swagger and built using Node.js"
    },
    "paths": {
...
        "/updateEmployeeDataByHR": {
            "put": {
                "tags": [
                    "TMS Employee Details"
                ],
                "operationId": "updateEmployeeDataByHR",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "updateEmployeeDataByHR",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateEmployee"
                          }
                    }
                   
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Contact"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
...
        "/updatesubskillcategoryforadmin": {
            "put": {
                "tags": [
                    "TMS Employee Details"
                ],
                "operationId": "updatesubskillcategoryforadmin",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "updatesubSkillMaster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updatesubSkillMaster"
                          }
                    }
                   
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Contact"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/deleteEmployeeResourceSkillsByHR": {
            "put": {
                "tags": [
                    "TMS Employee Details"
                ],
                "operationId": "deleteEmployeeResourceSkillsByHR",
                "consumes": [],
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "deleteEmployeeSkills",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/deleteResourceskills"
                          }
                    }
                   
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Contact"
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
        "Contact": {
            "type": "object",
            "properties": {
                "id": {
                    "format": "int32",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
...
        "updateEmployee": {
            "type": "object",
            "properties": {
                "FirstName": {
                "type": "string"
                },
                "LastName": {
                    "type": "string"
                },
                "DesignationId": {
                    "format": "int32",
                    "type": "integer"
                },
                "Password": {
                    "type": "string"
                },
                "QualificationId": {
                    "format": "int32",
                    "type": "integer"
                },
                "ContactNo": {
                    "type": "string"
                },
                "Gender": {
                    "type": "string"
                },
                "PriorExprience": {
                    "format":"float",
                    "type": "number"
                },
                "integraExperience": {
                    "format":"float",
                    "type": "number"
                },
                "DateOfBirth": {
                    "type": "string",
                    "format": "date"
                },
                "DateOfJoining": {
                    "type": "string",
                    "format": "date"
                },
                
                "ModifiedDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "RoleId": {
                    "format": "int32",
                    "type": "integer"
                },
                "PrimarySkill": {
                    "format": "int32",
                    "type": "integer"
                },
                "ReportingManager": {
                    "type": "string"
                },
                "IsReportingHead": {
                    "format": "int32",
                    "type": "integer"
                },
                "ModifiedBy": {
                    "type": "string"
                   
                },
                "Availability": {
                    "format": "int32",
                    "type": "integer"
                },
                "BuId":{
                    "format": "int32",
                    "type": "integer"
                },
                "EmployeeId": {
                    "type": "string"
                }
            
            }
        },
...
        "updatesubSkillMaster": {
            "type": "object",
            "properties": {
                "CategoryId": {
                    "format": "int32",
                    "type": "integer"
                },
                "SkillName": {
                    "type": "string"
                },
                "Description": {
                    "type": "string"
                },
                
                "date": {
                    "type": "string",
                    "format": "date-time"
                },
                "SkillId": {
                    "format": "int32",
                    "type": "integer"
                }
            }
        },
...
        "deleteResourceskills": {
            "type": "object",
            "properties": {
                "EmployeeSkillId": {
                    "format": "int32",
                    "type": "integer"
                },
                "UserId": {
                    "format": "int32",
                    "type": "integer"
                },
                "ModifiedDate": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    }
}